Move app logic from libs.rs to main.rs
