+++ src/lib.rs
@@ -3,13 +3,8 @@
 #![no_std]
 #![feature(llvm_asm)]
 
-extern crate alloc;
 extern crate rlibc;
-extern crate uefi_std as std;
 
-use alloc::collections::BTreeMap;
-use std::{mem, slice};
-use uefi::guid::Guid;
 use uefi::status::{Error, Result};
 
 unsafe fn smm_cmd(cmd: u8, subcmd: u8, arg: u32) -> u32 {
@@ -39,11 +34,11 @@ const SMMSTORE_CLEAR: u8 = 1;
 const SMMSTORE_READ: u8 = 2;
 const SMMSTORE_APPEND: u8 = 3;
 
-unsafe fn smmstore_clear() -> Result<()> {
+pub unsafe fn smmstore_clear() -> Result<()> {
     smmstore_cmd(SMMSTORE_CLEAR, 0)
 }
 
-unsafe fn smmstore_read(buf: &mut [u8]) -> Result<()> {
+pub unsafe fn smmstore_read(buf: &mut [u8]) -> Result<()> {
     #[repr(C)]
     struct Params {
         buf: u32,
@@ -56,7 +51,7 @@ unsafe fn smmstore_read(buf: &mut [u8]) -> Result<()> {
     smmstore_cmd(SMMSTORE_READ, &params as *const Params as u32)
 }
 
-unsafe fn smmstore_append(key: &[u8], val: &[u8]) -> Result<()> {
+pub unsafe fn smmstore_append(key: &[u8], val: &[u8]) -> Result<()> {
     #[repr(C)]
     struct Params {
         key: u32,
@@ -71,72 +66,4 @@ unsafe fn smmstore_append(key: &[u8], val: &[u8]) -> Result<()> {
         valsize: val.len() as u32
     };
     smmstore_cmd(SMMSTORE_APPEND, &params as *const Params as u32)
-}
-
-pub fn smmstore() -> Result<()> {
-    let mut data = [0; 0x40000];
-    let res = unsafe { smmstore_read(&mut data) };
-    // println!("Read {:?}", res);
-    res?;
-
-    let mut compact = BTreeMap::<&[u8], &[u8]>::new();
-
-    let mut i = 0;
-    let mut duplicates = 0;
-    while i + 8 <= data.len() {
-        let (keysz, valsz) = unsafe {
-            let ptr = data.as_ptr().add(i) as *const u32;
-            i += 8;
-            (*ptr as usize, *ptr.add(1) as usize)
-        };
-
-        // No more entries
-        if keysz == 0 || keysz == 0xffff_ffff {
-            break;
-        }
-
-        // Data too short
-        if i + keysz + valsz >= data.len() {
-            break;
-        }
-
-        unsafe {
-            let ptr = data.as_ptr().add(i);
-            let key = slice::from_raw_parts(
-                ptr,
-                keysz
-            );
-            let value = slice::from_raw_parts(
-                ptr.add(keysz),
-                valsz
-            );
-            if compact.insert(key, value).is_some() {
-                duplicates += 1;
-            }
-        }
-
-        i += keysz + valsz + 1;
-        i = (i + 3) & !3;
-    }
-
-    let rewrite = {
-        i >= data.len() / 2 ||
-        duplicates >= 16
-    };
-
-    if rewrite {
-        let res = unsafe { smmstore_clear() };
-        // println!("Clear {:?}", res);
-        res?;
-
-        for (key, value) in compact.iter() {
-            if key.len() > mem::size_of::<Guid>() && value.len() > 0 {
-                let res = unsafe { smmstore_append(&key, &value) };
-                // println!("Append {:?}", res);
-                res?;
-            }
-        }
-    }
-
-    Ok(())
 }
